+++ internal/app/service.ent/demo.srv.go
@@ -137,7 +137,7 @@ func (a *Demo) Delete(ctx context.Context, id string) error {
 }
 
 // UpdateStatus 更新状态
-func (a *Demo) UpdateStatus(ctx context.Context, id string, status int) error {
+func (a *Demo) UpdateStatus(ctx context.Context, id string, status int16) error {
 	oldItem, err := a.DemoModel.Get(ctx, id)
 	if err != nil {
 		return err
