+++ internal/app/service.ent/menu.srv.go
@@ -549,7 +549,7 @@ func (a *Menu) Delete(ctx context.Context, id string) error {
 }
 
 // UpdateStatus 更新状态
-func (a *Menu) UpdateStatus(ctx context.Context, id string, status int) error {
+func (a *Menu) UpdateStatus(ctx context.Context, id string, status int16) error {
 	oldItem, err := a.MenuModel.Get(ctx, id)
 	if err != nil {
 		return err
