+++ internal/schema/repo_ent/menu.repo.go
@@ -190,7 +190,7 @@ func (a *Menu) Delete(ctx context.Context, id string) error {
 }
 
 // UpdateStatus 更新状态
-func (a *Menu) UpdateStatus(ctx context.Context, id string, status int) error {
+func (a *Menu) UpdateStatus(ctx context.Context, id string, status int16) error {
 	_, err := a.EntCli.SysMenu.UpdateOneID(id).SetStatus(status).Save(ctx)
 	return errors.WithStack(err)
 }
