+++ internal/schema/repo_ent/role.repo.go
@@ -191,7 +191,7 @@ func (a *Role) Delete(ctx context.Context, id string) error {
 }
 
 // UpdateStatus 更新状态
-func (a *Role) UpdateStatus(ctx context.Context, id string, status int) error {
+func (a *Role) UpdateStatus(ctx context.Context, id string, status int16) error {
 	_, err1 := a.EntCli.SysRole.UpdateOneID(id).SetStatus(status).Save(ctx)
 	return errors.WithStack(err1)
 }
