+++ internal/schema/repo_ent/user.repo.go
@@ -186,7 +186,7 @@ func (a *User) Delete(ctx context.Context, id string) error {
 }
 
 // UpdateStatus 更新状态
-func (a *User) UpdateStatus(ctx context.Context, id string, status int) error {
+func (a *User) UpdateStatus(ctx context.Context, id string, status int16) error {
 	_, err := a.EntCli.SysUser.UpdateOneID(id).SetStatus(status).Save(ctx)
 	return errors.WithStack(err)
 }
